create and populate module for functions dealing with HOG and comp ids
